+++ CHANGES.md
@@ -3,6 +3,12 @@
 Fedify changelog
 ================
 
+Version 0.14.0
+--------------
+
+To be released.
+
+
 Version 0.13.0
 --------------